requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
openai-whisper
PyPDF2>=3.0.0
llama-index>=0.10.0
llama-index-readers-file>=0.1.0
pytesseract>=0.3.10
pillow>=10.0.0
pdf2image>=1.16.0 